# docs/server/openapi.v1.yaml
openapi: 3.1.0
info:
  title: SpeechKit Server API
  version: 1.0.0
  description: |
    Stable v1 HTTP and WebSocket contract for the SpeechKit Server.
    The canonical server API prefix is `/v1/`; `/api/v1/` remains
    available as a compatibility alias. These endpoints are covered by the same
    backwards-compatibility guarantees as the Framework's public
    `pkg/speechkit` packages — fields are additive across minor releases; no
    rename or removal without a version bump.

    Ships with the container image
    `ghcr.io/kombifyio/speechkit-server:v0.26.0` and later. Earlier images
    expose the same surface but were Server-internal pre-v0.26.

  contact:
    url: https://github.com/kombifyio/SpeechKit
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.speechkit.dev
    description: Replace with your deployment's URL.

# ── Auth ────────────────────────────────────────────────────────────────────
# Default server deployments require bearer auth. `auth_mode = "none"` is only
# intended for loopback/dev deployments and must not be used on public binds.
security:
  - {}
  - bearer: []
  - edgeHmac: []
paths:
...
  /v1/catalog/profiles:
    get:
      operationId: listCatalogProfiles
      summary: List provider profiles
      tags: [catalog]
      parameters:
        - name: mode
          in: query
          required: false
          schema: { type: string, enum: [dictation, assist, voiceagent, voice_agent] }
      responses:
        '200':
          description: Provider catalog.
          content:
            application/json:
              schema:
                type: object
                properties:
                  profiles:
                    type: array
                    items: { $ref: '#/components/schemas/ProviderProfile' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
  /v1/catalog/contracts:
    get:
      operationId: listCatalogContracts
      summary: List strict mode contracts
      tags: [catalog]
      responses:
        '200':
          description: Mode contracts.
          content:
            application/json:
              schema:
                type: object
                properties:
                  contracts:
                    type: array
                    items: { $ref: '#/components/schemas/ModeContract' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
  /v1/catalog/readiness:
    get:
      operationId: listCatalogReadiness
      summary: List provider readiness snapshots
      tags: [catalog]
      responses:
        '200':
          description: Readiness for every provider profile.
          content:
            application/json:
              schema:
                type: object
                properties:
                  readiness:
                    type: array
                    items: { $ref: '#/components/schemas/ProviderReadiness' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
  /v1/voiceagent/sessions:
    post:
      operationId: createVoiceAgentSession
      summary: Create a Voice Agent session and ticket
      description: |
        Creates a session record and returns a one-time HMAC ticket the
        client uses to upgrade the WebSocket. Tickets expire (default 30 s)
        and are bound to the session ID — they cannot be replayed against
        a different session.

        The WebSocket upgrade itself happens at
        `GET /v1/voiceagent/sessions/{id}/ws?ticket=...` and is documented
        separately because OpenAPI 3.1 has no first-class WebSocket
        primitive.

        When `[server.livekit]` is enabled, the response also includes a
        short-lived LiveKit join token for the same authenticated caller.
        Render-hosted LiveKit is the target media service for production;
        this API does not require a host migration. SpeechKit remains the
        token issuer and policy boundary.

        WebSocket control frames are JSON text messages. The first client
        frame must be `{"type":"start"}` and may include `persona_id`,
        `role_id`, `sequence_id`, `media_transport`, `voice`, `locale`,
        `model`, `activity_detection`, and `system_prompt_override`.
        `media_transport` defaults to `websocket` for backwards compatibility.
        Set `media_transport` to `livekit` to keep the WebSocket as the
        control channel while microphone and model audio move through the
        LiveKit room from this response. v1 LiveKit media is limited to native
        realtime PCM providers (Gemini/OpenAI); cascaded providers remain on
        WebSocket audio until explicit transcoding is added. During a
        workflow, clients send `{"type":"advance_step","reason":"host"}` to
        move to the next sequence step. The server emits `sequence_step`
        frames with `status` values `entered`, `completed`, and
        `sequence_completed`. Tool calls use `tool_call` and `tool_response`.
      tags: [voiceagent]
      responses:
        '201':
          description: Session created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreateSessionResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '409':
          description: Per-identity session limit exceeded.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503':
          description: Global session limit reached.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      operationId: listVoiceAgentSessions
      summary: List the caller's active sessions
      tags: [voiceagent]
      responses:
        '200':
          description: List returned (possibly empty).
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items: { $ref: '#/components/schemas/ListedSession' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: |
        Static deployment-wide token from `$SPEECHKIT_SERVER_TOKEN`. The
        identity attached to the request context is `service@default/internal`.
        Suitable for internal service-to-service callers, product backends, or
        automation services on a private network.
    edgeHmac:
      type: apiKey
      in: header
      name: X-Edge-Auth-Hmac
      description: |
        HMAC-SHA256 of `userID + "\n" + orgID + "\n" + plan + "\n" + role` using
        `$EDGE_AUTH_SECRET`. Must be sent alongside `X-Edge-User-Id`,
        `X-Edge-Org-Id`, `X-Edge-Plan`, and optional `X-Edge-Role`. Used
        when fronted by a Cloudflare Worker / reverse proxy that has
        already validated a user JWT.
...
  responses:
...
    Unauthenticated:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable identifier.
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    ProviderProfile:
      type: object
      properties:
        id: { type: string }
        mode: { type: string }
        name: { type: string }
        providerKind: { type: string }
        executionMode: { type: string }
        modelId: { type: string }
        source: { type: string }
        description: { type: string }
        license: { type: string }
        capabilities:
          type: array
          items: { type: string }
        adapterKind: { type: string }
        inferenceAllowed: { type: boolean }
        default: { type: boolean }
        recommended: { type: boolean }
        experimental: { type: boolean }
    ModeContract:
      type: object
      properties:
        mode: { type: string }
        intelligence: { type: string }
        input: { type: string }
        output: { type: string }
        allowed:
          type: array
          items: { type: string }
        forbidden:
          type: array
          items: { type: string }
    ProviderReadiness:
      type: object
      description: |
        Server-side readiness for a catalog profile. `active` means selected
        by server configuration; `default` is static catalog metadata.
        `ready` is independent of `active` and only reflects whether the
        profile could run if selected.
      properties:
        schemaVersion: { type: string }
        profileId: { type: string }
        mode: { type: string }
        providerKind: { type: string }
        executionMode: { type: string }
        modelId: { type: string }
        source: { type: string }
        active: { type: boolean }
        default: { type: boolean }
        configured: { type: boolean }
        credentialsReady: { type: boolean }
        runtimeReady: { type: boolean }
        capabilityReady: { type: boolean }
        ready: { type: boolean }
        missing:
          type: array
          items:
            type: string
            enum: [mode_disabled, provider_disabled, credentials, runtime]
...
    CreateSessionResponse:
      type: object
      properties:
        session_id: { type: string }
        ws_url:
          type: string
          format: uri
          description: |
            Use this exact URL (already includes the `?ticket=…` query
            param) to open the WebSocket. Tickets are HMAC-bound to the
            session id and expire after the configured TTL (default 30 s).
        ticket: { type: string }
        expires_at:
          type: string
          format: date-time
        livekit:
          $ref: '#/components/schemas/LiveKitJoinInfo'
    LiveKitJoinInfo:
      type: object
      properties:
        url:
          type: string
          format: uri
          description: LiveKit WebSocket URL, e.g. `wss://livekit.example.com`.
        room:
          type: string
          description: SpeechKit-managed LiveKit room name.
        token:
          type: string
          description: Short-lived LiveKit JWT scoped to this room and participant.
        token_expires_at:
          type: string
          format: date-time
        participant_identity:
          type: string
    ListedSession:
      type: object
      properties:
        session_id: { type: string }
        state:
          type: string
          enum: [pending_ws, active, closed]
        created_at:
          type: string
          format: date-time
        has_ws_client: { type: boolean }
...